KAFKA-14683 Migrate #testTopicsRegex to Mockito
